requirements.txt
azure-cognitiveservices-vision-face

msrest~=0.6.21
requests~=2.25.1
pillow
